feat : fix TodoController, TodoService * todo 등록,삭제,수정시 인증 요구 * todo 수정,삭제시 본인 확인 기능 추가 * 특정 유저 todo목록 조회 URL 변경 * TodoEntity User와 매핑

+++ src/main/java/com/example/todo/service/todo/TodoApiService.java
@@ -1,11 +1,15 @@
 package com.example.todo.service.todo;
 
+import com.example.todo.domain.entity.TeamEntity;
+import com.example.todo.domain.entity.user.User;
 import com.example.todo.domain.repository.TodoApiRepository;
+import com.example.todo.domain.repository.user.UserRepository;
 import com.example.todo.dto.ResponseDto;
 import com.example.todo.dto.todo.TodoApiDto;
 import com.example.todo.domain.entity.TodoApiEntity;
 import com.example.todo.exception.ErrorCode;
 import com.example.todo.exception.TodoAppException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -21,7 +25,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class TodoApiService {
     private final TodoApiRepository todoApiRepository;
-
+    private final UserRepository userRepository;
     //해당 to do가  존재하는지 확인하는 메소드
     public TodoApiEntity getTodoById(Long id) {
         Optional<TodoApiEntity> optionalTodoApiEntity = todoApiRepository.findById(id);
@@ -29,11 +33,20 @@ public class TodoApiService {
             throw new TodoAppException(ErrorCode.NOT_FOUND_TODO);
         return optionalTodoApiEntity.get();
     }
-
+    //해당 유저가 존재하는지 확인
+    public User getUserById(Long userId) {
+        Optional<User> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isEmpty())
+            throw new TodoAppException(ErrorCode.NOT_FOUND_USER);
+        return optionalUser.get();
+    }
+@Transactional
     //to do 등록
-    public ResponseDto createTodo(TodoApiDto todoApiDto) {
-
+    public ResponseDto createTodo(Long userId, TodoApiDto todoApiDto) {
         TodoApiEntity todoApiEntity = new TodoApiEntity();
+        User user = getUserById(userId);
+
+        todoApiEntity.setUser(user);
         todoApiEntity.setTitle(todoApiDto.getTitle());
         todoApiEntity.setContent(todoApiDto.getContent());
         todoApiEntity.setStartDate(todoApiDto.getStartDate());
@@ -72,15 +85,22 @@ public class TodoApiService {
         optionalTodoApiEntity.get();
         //존재한다면 조회하기
         Pageable pageable = PageRequest.of(page, limit);
-        Page<TodoApiEntity> todoApiEntities = todoApiRepository.findAll(pageable);
+        Page<TodoApiEntity> todoApiEntities = todoApiRepository.findByUserId(userId, pageable);
         Page<TodoApiDto> todoApiDtos = todoApiEntities.map(TodoApiDto::fromEntity);
         return todoApiDtos;
     }
 
     //to do 수정
-    public ResponseDto updateTodo(Long todoId, TodoApiDto todoApiDto) {
+    public ResponseDto updateTodo(Long userId, Long todoId, TodoApiDto todoApiDto) {
         TodoApiEntity todoApiEntity = getTodoById(todoId);
+        User user = getUserById(userId);
 
+        // To do 작성자인지 확인
+        if (!todoApiEntity.getUser().getId().equals(userId)) {
+            throw new TodoAppException(ErrorCode.NOT_MATCH_USERID);
+        }
+
+        todoApiEntity.setUser(user);
         todoApiEntity.setTitle(todoApiDto.getTitle());
         todoApiEntity.setContent(todoApiDto.getContent());
         todoApiEntity.setStartDate(todoApiDto.getDueDate());
@@ -101,8 +121,14 @@ public class TodoApiService {
     }
 
     //to do 삭제
-    public ResponseDto deleteTodo(Long todoId) {
+    public ResponseDto deleteTodo(Long userId, Long todoId) {
         TodoApiEntity todoApiEntity = getTodoById(todoId);
+
+    // To do 작성자인지 확인
+        if (!todoApiEntity.getUser().getId().equals(userId)) {
+            throw new TodoAppException(ErrorCode.NOT_MATCH_USERID);
+        }
+
         todoApiRepository.deleteById(todoApiEntity.getId());
         return new ResponseDto("Todo가 삭제되었습니다.");
     }
